Combined DialogAgentMqtt into its base class
